+++ SplunkAppForWazuh/appserver/static/js/config/routes/security-states.js
@@ -128,6 +128,26 @@ define(['../module'], function(module) {
                 return false
               }
             }
+          ],
+          roles: [
+            '$requestService',
+            async $requestService => {
+              try {
+                return await $requestService.apiReq('/security/roles')
+              } catch (err) {
+                return { error: 'Cannot fetch roles from API' }
+              }
+            }
+          ],
+          splunkUsers: [
+            '$splunkUsers',
+            async $splunkUsers => {
+              try {
+                return await $splunkUsers.getInternalUsers()
+              } catch (err) {
+                return { error: 'Cannot fetch splunk users from API' }
+              }
+            }
           ]
         }
       })