+++ SplunkAppForWazuh/bin/get_agents_status.py
@@ -27,8 +27,16 @@ db = database()
 logger = log()
 wztoken = wazuhtoken()
 
+
 def get_apis():
-    """Obtain the list of APIs."""
+    """
+    Obtain the list of APIs.
+    
+    Returns
+    -------
+    string
+        a JSON string containing the saved APIs information.
+    """
     try:
         logger.debug("bin.get_agents_status: Getting APIs.")
         session_key = getSplunkSessionKey()
@@ -43,7 +51,8 @@ def check_status():
     try:
         logger.debug("bin.get_agents_status: Checking agents status.")
         apis = get_apis()
-        apis = jsonbak.loads(apis) #get_apis() returns a JSON string, it needs to be converted as a dictionary
+        # get_apis() returns a JSON string, it needs to be converted as a dictionary
+        apis = jsonbak.loads(apis)
         date = str(datetime.datetime.utcnow())[:-7]
         # obtains
         for api in apis:
@@ -54,31 +63,32 @@ def check_status():
             agent_list = {}
             url = str(opt_base_url) + ":" + str(opt_base_port)
             auth = requestsbak.auth.HTTPBasicAuth(opt_username, opt_password)
-            wazuh_token = wztoken.get_auth_token(url,auth)
+            wazuh_token = wztoken.get_auth_token(url, auth)
             verify = False
             agents_url_total_items = url + '/agents?limit=1&q=id!=000'
             try:
                 request_agents = requestsbak.get(
                     agents_url_total_items,
-                     headers = {'Authorization': f'Bearer {wazuh_token}'}, timeout=1,
+                    headers={'Authorization': f'Bearer {wazuh_token}'}, timeout=1,
                     verify=verify).json()
                 total_items = request_agents["data"]["total_affected_items"]
                 limit = 500
                 offset = 0
                 params = {}
                 params['q'] = 'id!=000'
-                
+
                 while offset < total_items:
                     agents_url = url + \
-                        '/agents?select=id,ip,manager,status&offset='+str(offset)+'&limit='+str(limit)
+                        '/agents?select=id,ip,manager,status&offset=' + \
+                        str(offset)+'&limit='+str(limit)
                     request_agents = requestsbak.get(
-                        agents_url, headers = {'Authorization': f'Bearer {wazuh_token}'}, timeout=1, verify=verify).json()
+                        agents_url, headers={'Authorization': f'Bearer {wazuh_token}'}, timeout=1, verify=verify).json()
 
                     agent_list = request_agents["data"]["affected_items"]
                     final_url_cluster = url + '/cluster/status'
                     request_cluster_status = requestsbak.get(
                         final_url_cluster,
-                        headers = {'Authorization': f'Bearer {wazuh_token}'},
+                        headers={'Authorization': f'Bearer {wazuh_token}'},
                         timeout=1,
                         verify=verify).json()
                     cluster_status = request_cluster_status["data"]["enabled"]
@@ -87,7 +97,7 @@ def check_status():
                         request_cluster_name = requestsbak.get(
                             final_url_cluster_name,
                             timeout=1,
-                            headers = {'Authorization': f'Bearer {wazuh_token}'},
+                            headers={'Authorization': f'Bearer {wazuh_token}'},
                             verify=verify).json()
                     offset = offset + limit
                     for item in agent_list:
@@ -101,17 +111,23 @@ def check_status():
                         item["timestamp"] = date
                         print(jsonbak.dumps(item))
             except Exception as e:
-                logger.error("agents-status: Error requesting agents status: %s" % str(e))
+                logger.error(
+                    "agents-status: Error requesting agents status: %s" % str(e))
                 pass
     except Exception as e:
-        logger.error("agents-status: Error requesting agents status: %s" % str(e))
+        logger.error(
+            "agents-status: Error requesting agents status: %s" % str(e))
         pass
 
 
 def getSplunkSessionKey():
-    """Get the session key, it needs to configure in the inputs.conf that executes this script the following parameter: passAuth = splunk-system-user"""
+    """
+    Get the session key, it needs to configure in the inputs.conf that 
+    executes this script the following parameter: passAuth = splunk-system-user
+    """
     logger.debug("bin.get_agents_status: Getting Splunk session key.")
     session_key = sys.stdin.readline().strip()
     return session_key
+
 
 check_status()